refactor: implement interface for File Repository and update files
