+++ morpheus-server/app/repository/user_repository.py
@@ -3,7 +3,7 @@ from typing import Union, List
 from sqlalchemy.orm import Session
 
 from app.models.models import User
-from app.repository.files_repository import IMAGES_BUCKET
+from app.repository.files.s3_files_repository import IMAGES_BUCKET
 
 
 class UserRepository:
