+++ morpheus-server/scripts/models/s3.py
@@ -8,14 +8,14 @@ from config import Target, config_file, download_model
 from utils import load_config_from_file
 
 sys.path.append(".")
-from app.repository.files_repository import ModelFileRepository  # noqa: E402
+from app.repository.files.s3_models_repository import S3ModelFileRepository  # noqa: E402
 
 app = typer.Typer(help="subcommand to manage models in S3:  register/list/delete")
 
 
 def register_model_in_s3(output_path: str) -> None:
-    model_repository = ModelFileRepository()
-    is_in_s3 = model_repository.model_exist_and_not_empty_in_s3(output_path)
+    model_repository = S3ModelFileRepository()
+    is_in_s3 = model_repository.files_exists(output_path)
 
     if is_in_s3:
         print("Model already exists in s3")
@@ -23,20 +23,20 @@ def register_model_in_s3(output_path: str) -> None:
 
     list_of_content = glob(f"./tmp/{output_path}/**/*", recursive=True)
     list_of_files = [file for file in list_of_content if Path(file).is_file()]
-    model_repository.upload_diffuser_model_to_s3(list_of_files)
+    model_repository.upload_files(list_of_files)
     print("Model uploaded to the S3 bucket")
 
 
 def delete_model_from_s3(name):
-    model_repository = ModelFileRepository()
-    model_repository.delete_model_from_s3(name.replace(" ", "_"))
+    model_repository = S3ModelFileRepository()
+    model_repository.delete_file(name.replace(" ", "_"))
     print("Model deleted from S3 bucket")
 
 
 @app.command("list")
 def list(folder: str = typer.Argument("")) -> None:
-    model_repository = ModelFileRepository()
-    model_repository.list_s3_content(folder=folder)
+    model_repository = S3ModelFileRepository()
+    model_repository.list_content(folder=folder)
 
 
 @app.command("register")
