+++ morpheus-server/secrets.env.dist
@@ -21,6 +21,8 @@ AWS_SECRET_ACCESS_KEY=
 
 # S3 BUCKETS ID
 #------------------------
+# Bucket type depending on the cloud provider
+BUCKET_TYPE=
 # Bucket name where images are stored
 IMAGES_BUCKET=
 # Bucket name where models are stored
